+++ src/game/tests.rs
@@ -415,6 +415,7 @@ mod vintage_cube_fable_of_the_mirror_breaker;
 mod vintage_cube_fabled_passage;
 mod vintage_cube_faerie_mastermind;
 mod vintage_cube_fanatic_of_rhonas;
+mod vintage_cube_fastbond;
 mod vintage_cube_fatal_push;
 mod vintage_cube_fear_of_missing_out;
 mod vintage_cube_field_of_the_dead;